Add ability to debug EBeans instantiation
Now it's possible to set a hook on the @EBean-annotated class
instantiation. FactoryHook.setHook() may be used for this purpose.
The installed hook is first called on bean creation if the bean is not a
singleton and not already cleated. In either case the hook's
onInstanceRequested() also will be called in getInstance_(), No matter,
was new instance creaded or not.

Both hook's methods are called after object initialization so everything
should be injected at this moment.