Instantiate Beans only on UI thread
Now the .getInstance_() method of the generated subclass of the
@EBean-annotated class creates the Bean instance only on UI thread.
If it is called on UI thread then everything works as before. If it is
called on background thread then the task for creating the instance is
posted to the UI thread Handler and the method waits while the task is
finishes.