Optimize obtaining of the existing Singleton Beans
No need to wait for UI thread Handler to get a 'Singleton' Bean instance
if it has been already instantiated and initialized. In this case it can
be returned immediately. Instantiation/initializing must be atomic from
the viewpoint of checking for existance of the instance so both blocks
will be synchronized.